chore(wms): move warehouse model to domain package
